sistema_distribucion/requirements.txt
joblib==1.3.2
numpy==1.26.1
paho-mqtt==1.6.1
pandas==2.1.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
scikit-learn==1.3.0
scipy==1.11.3
six==1.16.0
threadpoolctl==3.2.0
tzdata==2023.3
